requirements.txt
absl-py==1.4.0
alembic==1.11.2
altair==5.0.1
ansi2html==1.8.0
anyio==3.7.1
appnope==0.1.3
argcomplete==3.2.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.16.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.1.0
blis==0.7.10
bokeh==3.2.1
Boruta==0.3
boto3==1.28.20
botocore==1.31.20
branca==0.6.0
brewer2mpl==1.4.1
Brotli==1.0.9
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.1
calplot==0.1.7.5
catalogue==2.0.9
catboost==1.2
category-encoders==2.6.2
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
cleo==2.1.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cmaes==0.10.0
colorlog==6.7.0
comm==0.1.4
commonmark==0.9.1
confection==0.1.1
contourpy==1.1.0
cookiecutter==2.5.0
crashtest==0.4.1
croniter==1.4.1
cycler==0.11.0
cymem==2.0.7
Cython==3.0.0
dash==2.11.1
dash-bootstrap-components==1.3.1
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.8.0
databricks-cli==0.17.7
DateTime==5.2
deap==1.4.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
dm-tree==0.1.8
docker==6.1.3
dtale==3.3.0
dulwich==0.21.7
eli5==0.13.0
ensure==1.0.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.8.0.post1
fastai==2.7.12
fastapi==0.104.1
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.18.0
fastprogress==1.0.3
featuretools==1.27.0
filelock==3.12.2
Fiona==1.9.4.post1
Flask==2.2.5
Flask-Compress==1.13
Flask-Cors==4.0.0
flask-ngrok==0.0.25
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.42.0
fqdn==1.5.1
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
gensim==4.3.1
geocoder==1.38.1
geographiclib==2.0
geopandas==0.13.2
geopy==2.4.1
ggplot==0.11.5
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.56.2
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5netcdf==1.2.0
h5py==3.9.0
hiredis==2.3.2
holidays==0.29
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
hyperopt==0.2.7
idna==3.4
ImageHash==4.3.1
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
install==1.3.5
installer==0.7.0
ipykernel==6.25.0
ipython==8.14.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
json5==0.9.24
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==2.13.1
keyring==24.3.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
libq==0.5.0
lime==0.2.0.1
llvmlite==0.40.1
locket==1.0.0
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mitdeeplearning==0.3.0
mlflow==2.2.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multimethod==1.9.1
murmurhash==1.0.9
nanoid==2.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.2.0
orjson==3.10.3
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.3
partd==1.4.0
pathlib==1.0.1
pathy==0.10.2
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.0
pipx==1.4.3
pkginfo==1.10.0
platformdirs==3.10.0
plotly==5.15.0
pmdarima==2.0.4
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
postgres==4.0
preshed==3.0.8
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyaml==23.9.6
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
pymc3==3.11.4
pyparsing==3.0.9
pyproj==3.6.0
pyproject_hooks==1.0.0
pyspark==3.4.1
python-box==6.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
querystring-parser==1.2.4
rapidfuzz==3.7.0
ratelim==0.1.6
redis==4.6.0
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.0
scikit-optimize==0.9.0
scipy==1.11.1
seaborn==0.13.2
semver==3.0.1
Send2Trash==1.8.2
shap==0.42.1
shapely==2.0.1
shellingham==1.5.4
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.19
sqlparse==0.4.4
squarify==0.4.3
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
stopit==1.1.2
strsimpy==0.2.1
sweetviz==2.1.4
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-probability==0.21.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
Theano-PyMC==1.1.2
thinc==8.1.10
threadpoolctl==3.2.0
tifffile==2023.7.18
tinycss2==1.2.1
tokenizers==0.19.1
tomlkit==0.12.4
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
TPOT==0.12.0
tqdm==4.65.0
traitlets==5.9.0
# Editable install with no version control (transformers==4.42.0.dev0)
-e /Users/main/Desktop/transformers
trove-classifiers==2024.3.25
typer==0.9.0
types-python-dateutil==2.8.19.20240106
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2023.3
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.16
userpath==1.9.1
utm==0.7.0
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.25.0
visions==0.7.4
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
Werkzeug==2.2.3
widgetsnbextension==4.0.9
woodwork==0.25.1
wordcloud==1.9.2
wrapt==1.15.0
xarray==2023.7.0
xarray-einstats==0.6.0
xattr==1.1.0
xgboost==1.7.6
xlrd==2.0.1
xyzservices==2023.7.0
yellowbrick==1.5
zipp==3.16.2
zope.interface==6.0
